Multi Layer Pattern Refactor
