+++ SharpBank.API/Controllers/AccountsController.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using SharpBank.API.Services.Interfaces;
+using SharpBank.Services.Interfaces;
 using SharpBank.Models;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
