+++ SharpBank.API/Controllers/BanksController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using SharpBank.API.Services.Interfaces;
 using SharpBank.Models;
+using SharpBank.Services.Interfaces;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
