--- SharpBank.API/Services/Interfaces/ITransactionService.cs
@@ -1,14 +1,0 @@
-﻿using SharpBank.Models;
-
-namespace SharpBank.API.Services.Interfaces
-{
-    public interface ITransactionService
-    {
-        public IEnumerable<Transaction> GetTransactions();
-        public Transaction GetTransactionById(Guid Id);
-
-        public Transaction Create(Transaction transaction);
-        public Transaction Update(Transaction transaction);
-        public Transaction Delete(Guid Id);
-    }
-}
