+++ SharpBank.API/appsettings.json
@@ -1,5 +1,5 @@
 {
-  "ConnectionStrings": { "SharpDBCS": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=SharpBankDB;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False" },
+  "ConnectionStrings": { "SharpDBCS": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=SharpBank;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False" },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
