+++ SharpBank.Services/AccountService.cs
@@ -1,4 +1,6 @@
-﻿using SharpBank.Models;
+﻿using Microsoft.EntityFrameworkCore;
+using SharpBank.Data;
+using SharpBank.Models;
 using SharpBank.Services.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -10,38 +12,51 @@ namespace SharpBank.Services
 {
     public class AccountService : IAccountService
     {
+        private readonly AppDbContext appDbContext;
+
         public AccountService(AppDbContext appDbContext)
         {
-
+            this.appDbContext = appDbContext;
         }
-        public Account Authenticate(long accountId, string password)
+        public Account Authenticate(Guid accountId, string password)
         {
             throw new NotImplementedException();
         }
 
         public Account Create(Account account)
         {
-            throw new NotImplementedException();
+            appDbContext.Accounts.Add(account);
+            appDbContext.SaveChanges();
+            return account;
         }
 
-        public Account Delete(long accountId)
+        public Account Delete(Guid accountId)
         {
             throw new NotImplementedException();
         }
 
-        public Account GetAccountById(long accountId)
+        public Account GetAccountById(Guid accountId)
         {
-            throw new NotImplementedException();
+            var res = appDbContext.Accounts.FirstOrDefault(a => a.AccountId == accountId);
+            return res;
         }
 
         public IEnumerable<Account> GetAccounts()
         {
-            throw new NotImplementedException();
+            return appDbContext.Accounts
+                .Include(a => a.Funds)
+                .ThenInclude(f => f.Wallets)
+                .Include(a => a.CreditTransactions)
+                .Include(a => a.DebitTransactions)
+                .ToList();
         }
 
         public Account Update(Account account)
         {
-            throw new NotImplementedException();
+            appDbContext.Accounts.Attach(account);
+            appDbContext.SaveChanges();
+            return account;
         }
     }
+
 }
