+++ SharpBank.Services/Interfaces/IAccountService.cs
@@ -7,5 +7,16 @@ using System.Threading.Tasks;
 
 namespace SharpBank.Services.Interfaces
 {
-    
+    public interface IAccountService
+    {
+        public Account Authenticate(Guid accountId, string password);
+        public IEnumerable<Account> GetAccounts();
+        public Account GetAccountById(Guid accountId);
+
+        public Account Create(Account account);
+        public Account Update(Account account);
+        public Account Delete(Guid accountId);
+
+
+    }
 }
